Add error-on-s3-exception option
This adds a new boolean option named `error-on-s3-exception` to be able
to mark the job as failed if the cache could not be restored.

This can be combined with a retry-action so we can use it in some situations
where we want to retry the job when the cache could not be restored.